Rename ref.h -> default.h for x86_64 native arithmetic backend
